+++ CHANGELOG.md
@@ -19,11 +19,13 @@ finished before its allotted execution timeout.
 - Fixed a bug where an event could erroneously be shown as silenced.
 - Properly log errors whenever a check request can't be published.
 - Fixed some build tags for tests using etcd stores.
+- Keepalive monitors now get updated with changes to a keepalive timeout.
 - Prevent tests timeout in queue package
 
 ### Changed
 - Queues are now durable.
 - Refactoring of the check scheduling integration tests.
+- CLI resource delete confirmation is now `(y/N)`.
 
 ## [2.0.0-alpha.15] - 2018-01-30
 ### Added